Correcting the code 0.5

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,13 +6,13 @@
  */
 int main(void)
 {
-	unsigned long n, a = 1, b = 1, s;
+	long long int  n, a = 1, b = 1, s;
 
-	printf("%lu", a);
+	printf("%lld", a);
 	for (n = 1; n <= 98; ++n)
 	{
 		s = a + b;
-		printf(", %lu", s);
+		printf(", %lld", s);
 		a = b;
 		b = s;
 	}
